Lesson 1-3

+++ src/lesson1/CanSwim.java
@@ -1,5 +1,5 @@
 package lesson1;
 
 public interface CanSwim {
-    public double swim(Pool pool);
+    public int swim(Pool pool);
 }
